+++ src/View/Content.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\View;
+
+final class Content
+{
+    /**
+     * @param callable(mixed, array<string, mixed>): string|array<string> $contentExtractor
+     */
+    public function __construct(
+        public readonly mixed $contentExtractor,
+        public readonly ?string $formatter = null
+    ) {
+    }
+}